requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.1
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djlint==1.36.4
EditorConfig==0.17.0
jsbeautifier==1.15.1
json5==0.10.0
pathspec==0.12.1
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
